requirements.txt

async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
discordbot.py==0.2.3a3
idna==2.10
multidict==5.1.0
psutil==5.8.0
pycodestyle==2.6.0
python-dotenv==0.15.0
requests==2.25.1
soupsieve==2.2
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
